UI-code/requirements.txt
django
djangorestframework
django-rest-auth
coreapi
django-bootstrap4
django-bootstrap-datepicker-plus